ExcelFeedback/requirements.txt
# Claude API
anthropic>=0.40.0

# PDF Processing
pdfplumber>=0.10.0
pikepdf>=8.0.0

# Excel Generation
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
